requirements.txt
addict==2.4.0
fiftyone==0.23.5
ftfy==6.1.3
graspnetAPI==1.2.10
huggingface_hub==0.17.3
ipdb==0.13.13
jsonlines==4.0.0
matplotlib==3.3.3
numpy==1.23.5
open3d==0.15.2
open3d_plus==0.1.0
opencv_python==4.4.0.46
Pillow==9.2.0
pybullet==3.0.7
PyYAML==6.0
regex==2023.10.3
Requests==2.31.0
scipy==1.12.0
setuptools==68.0.0
supervision==0.6.0
termcolor==1.1.0
timm==0.9.8
tqdm==4.57.0
transformers==4.34.1
typer==0.9.0
yapf==0.40.1
